@@ -9,10 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@capacitor-firebase/analytics": "^7.0.0",
-    "@capacitor-firebase/app": "^7.0.0",
-    "@capacitor-firebase/crashlytics": "^7.0.0",
-    "@capacitor-firebase/messaging": "^7.0.0",
+    "@capacitor-firebase/analytics": "^7.1.0",
+    "@capacitor-firebase/app": "^7.1.0",
+    "@capacitor-firebase/crashlytics": "^7.1.0",
+    "@capacitor-firebase/messaging": "^7.1.0",
     "@capacitor/android": "^7.0.1",
     "@capacitor/app": "^7.0.0",
     "@capacitor/core": "^7.0.1",
@@ -23,8 +23,8 @@
     "cordova-plugin-apkupdater": "^5.0.1",
     "dayjs": "^1.11.13",
     "firebase": "^11.4.0",
-    "next": "15.2.0",
-    "next-themes": "^0.4.4",
+    "next": "15.2.1",
+    "next-themes": "^0.4.5",
     "react": "19.0.0",
     "react-bootstrap-icons": "^1.11.5",
     "react-dom": "19.0.0",
@@ -35,12 +35,12 @@
   "devDependencies": {
     "@capacitor/assets": "^3.0.5",
     "@capacitor/cli": "^7.0.1",
-    "@tailwindcss/postcss": "^4.0.9",
+    "@tailwindcss/postcss": "^4.0.12",
     "daisyui": "^5.0.0",
-    "eslint": "^8.57.1",
-    "eslint-config-next": "15.2.0",
+    "eslint": "^9.22.0",
+    "eslint-config-next": "15.2.1",
     "postcss": "^8.4.47",
-    "tailwindcss": "^4.0.9"
+    "tailwindcss": "^4.0.12"
   },
   "packageManager": "yarn@4.6.0"
 }
