@@ -1,4 +1,5 @@
-import Layout from "@/components/layout";
+"use client";
+
 import { useAttendance, useAttendancesTypes, useLessons } from "@/lib/lessons";
 import { useSubjects, useTeachers } from "@/lib/school";
 import { upperFirst, groupAbsences } from "@/lib/utils";
@@ -53,7 +54,7 @@ const Attendance = () => {
   );
 
   return (
-    <Layout>
+    <div className="flex flex-col">
       <dialog
         id="attendanceDetails"
         className="modal modal-bottom sm:modal-middle"
@@ -397,6 +398,24 @@ const Attendance = () => {
       </dialog>
       <span className="text-3xl font-semibold mb-4">Attendance</span>
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 mt-4">
+        {!subjectsLoading && !subjectsError && (
+          <div role="alert" class="alert border border-base-300">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              fill="none"
+              viewBox="0 0 24 24"
+              class="stroke-info h-6 w-6 shrink-0"
+            >
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
+              ></path>
+            </svg>
+            <span>You can click on a subject to see more details.</span>
+          </div>
+        )}
         {!subjectsLoading && !subjectsError ? (
           subjectsData && subjectsData.length ? (
             subjectsData.map((subject) => {
@@ -519,7 +538,7 @@ const Attendance = () => {
           )}
         </div>
       </div>
-    </Layout>
+    </div>
   );
 };
 export default Attendance;
