@@ -1,5 +1,5 @@
-import { CapacitorCookies } from "@capacitor/core";
-import Layout from "@/components/layout";
+"use client";
+
 import { useState } from "react";
 
 const Developer = () => {
@@ -7,7 +7,7 @@ const Developer = () => {
   const [response, setResponse] = useState(null);
 
   return (
-    <Layout>
+    <div className="flex flex-col">
       <div className="flex flex-col">
         <span className="text-3xl font-semibold">Developer tools</span>
         <span className="text-lg">
@@ -101,7 +101,7 @@ const Developer = () => {
           </pre>
         )}
       </div>
-    </Layout>
+    </div>
   );
 };
 export default Developer;
