@@ -1,9 +1,10 @@
+"use client";
+
 import { Preferences } from "@capacitor/preferences";
 
 import { useTheme } from "next-themes";
 import { useEffect, useState } from "react";
 
-import Layout from "@/components/layout";
 import { upperFirst } from "@/lib/utils";
 
 const Settings = () => {
@@ -61,7 +62,7 @@ const Settings = () => {
   }, []);
 
   return (
-    <Layout>
+    <div className="flex flex-col">
       <dialog id="cacheClear" className="modal modal-bottom sm:modal-middle">
         <div className="modal-box">
           <h3 className="text-lg font-bold">Success!</h3>
@@ -249,7 +250,7 @@ const Settings = () => {
           </div>
         </div>
       </div>
-    </Layout>
+    </div>
   );
 };
 export default Settings;
