@@ -1,4 +1,5 @@
-import Layout from "@/components/layout";
+"use client";
+
 import { useTeachers } from "@/lib/school";
 import { useTeacherAbsences } from "@/lib/timetable";
 import { sortTeacherAbsences } from "@/lib/utils";
@@ -22,7 +23,7 @@ const TeacherAbsences = () => {
   );
 
   return (
-    <Layout>
+    <div className="flex flex-col">
       <span className="text-3xl font-semibold mb-4">Teacher absences</span>
       <div className="grid grid-cols-6 gap-2 mt-4">
         {!teacherAbsencesError &&
@@ -95,7 +96,7 @@ const TeacherAbsences = () => {
               ></div>
             ))}
       </div>
-    </Layout>
+    </div>
   );
 };
 export default TeacherAbsences;
